Use sp_trie layout impl instead
