+++ src/components/ndn-lib/src/path_object_map/mod.rs
@@ -1,6 +1,7 @@
 mod object_map;
 mod hash;
 mod storage;
+mod layout;
 
 pub use object_map::*;
 pub use hash::*;
