+++ src/components/ndn-lib/src/path_object_map/test.rs
@@ -47,17 +47,23 @@ fn generate_key_value_pairs(seed: &str, count: usize) -> Vec<(String, ObjId, Vec
 }
 
 #[test]
-async fn test_object_map() {
+async fn test_path_object_map() {
+    println!("Test object map");
     let key_pairs = generate_key_value_pairs("test", 100);
+    println!("Key pairs generated");
     let mut obj_map = PathObjectMap::new(HashMethod::Keccak256).await;
+    println!("Object map created");
 
     let count = 100;
     for i in 0..count {
+        println!("Test object map: {}/{}", i, count);
         let key = key_pairs[i].0.as_ref();
         let obj_id = key_pairs[i].1.clone();
         let meta = key_pairs[i].2.clone();
 
         obj_map.put_object(key, obj_id.clone(), Some(meta.clone())).await.unwrap();
+
+        println!("Put object success: {}", key);
 
         // Test get object
         let ret = obj_map.get_object(key).await.unwrap();
@@ -68,15 +74,21 @@ async fn test_object_map() {
         assert_eq!(ret.obj_id, obj_id);
         assert_eq!(ret.meta, Some(meta.clone()));
 
+        println!("Get object success: {}", key);
+
         // Test exist
         let ret = obj_map.is_object_exist(&key).await.unwrap();
         assert_eq!(ret, true);
+
+        println!("Object exist: {}", key);
 
         // Test remove
         if i % 2 == 0 {
             let ret = obj_map.remove_object(&key).await.unwrap().unwrap();
             assert_eq!(ret.0, obj_id);
             assert_eq!(ret.1, Some(meta));
+
+            println!("Remove object success: {}", key);
         }
     }
 }
